The very first version
This is just an initial program more for the programmer to actually understand how the boards will be reflected in the console than to create a functional and dynamic piece of code. I wrote a couple of years ago a minesweeper board program when I was learning Python and the beginning was similar. If you want to learn how to write your minesweeper game, be welcome to repeat my steps or use them as suggestions for your programming process. More refactorings and versions of this program soon to come!